#37 small changes
