+++ Code/GeorgiaLibrarySystem-/Tests/DbContextMock.cs
@@ -1,0 +1,21 @@
+﻿using System.Collections.Generic;
+using System.Data.Entity;
+using System.Linq;
+using Moq;
+
+namespace Tests
+{
+    public static class DbContextMock{
+        public static Mock<DbSet<T>> CreateDbSetMock<T>(IEnumerable<T> data) where T : class
+        {
+            var queryable = data.AsQueryable();
+            var mock = new Mock<DbSet<T>>();
+            mock.As<IQueryable<T>>().Setup(m => m.Provider).Returns(queryable.Provider);
+            mock.As<IQueryable<T>>().Setup(m => m.Expression).Returns(queryable.Expression);
+            mock.As<IQueryable<T>>().Setup(m => m.ElementType).Returns(queryable.ElementType);
+            mock.As<IQueryable<T>>().Setup(m => m.GetEnumerator()).Returns(() => queryable.GetEnumerator());
+ 
+            return mock;
+        }
+    }
+}
