+++ Code/GeorgiaLibrarySystem-/Tests/Tests.csproj
@@ -82,6 +82,7 @@
   <ItemGroup>
     <Compile Include="IntegrationTest\LendingTest.cs" />
     <Compile Include="IntegrationTest\LoginTest.cs" />
+    <Compile Include="DbContextMock.cs" />
     <Compile Include="UnitTest\LoginTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="UnitTest\LendingTest.cs" />
